add 3rd texture to list test adding textures at each world coordinate for 'load testing'

+++ src/main/resources/basic.vert
@@ -2,7 +2,7 @@
 
 in vec2 position;
 in vec2 texCoord;
-in int texIndex;
+in float texIndex;
 
 out vec3 textureCoord;
